+++ src/elements/simple-input/src/index.tsx
@@ -1,7 +1,10 @@
 /** @jsx jsx */
 
-import React, { useState } from 'react'
-import { jsx } from 'theme-ui'
+import React, { RefObject, useEffect, useState } from 'react'
+import { jsx, useThemeUI } from 'theme-ui'
+import get from '@uswitch/trustyle-utils.get'
+
+type InputSize = 'sm' | 'base' | undefined
 
 export interface Props extends React.InputHTMLAttributes<HTMLInputElement> {
   hasError?: boolean | undefined
@@ -12,6 +15,7 @@ export interface Props extends React.InputHTMLAttributes<HTMLInputElement> {
   defaultValue?: string
   prefix?: string
   suffix?: string
+  inputSize?: InputSize
 }
 
 export const SimpleInput = React.forwardRef(
@@ -22,14 +26,38 @@ export const SimpleInput = React.forwardRef(
       onChange,
       defaultValue,
       prefix,
-      suffix
+      suffix,
+      inputSize
     } = props
 
     const [hasFocus, setHasFocus] = useState(false)
+    const [prefixWidth, setPrefixWidth] = useState(0)
+    const [suffixWidth, setSuffixWidth] = useState(0)
+    const prefixRef: RefObject<HTMLDivElement> = React.createRef()
+    const suffixRef: RefObject<HTMLDivElement> = React.createRef()
+
+    useEffect(() => {
+      setPrefixWidth(prefixRef?.current?.offsetWidth || 0)
+      setSuffixWidth(suffixRef?.current?.offsetWidth || 0)
+    }, [prefixRef, suffixRef])
+
+    const setWidth = (inputSize: InputSize) => {
+      const { theme }: any = useThemeUI()
+
+      const widthLookup: { [key: string]: any } = {
+        sm: get(theme, 'elements.simple-input.inputWidth.small'),
+        base: '100%'
+      }
+
+      return inputSize
+        ? widthLookup[inputSize] + prefixWidth + suffixWidth
+        : widthLookup.base
+    }
 
     return (
       <div
         sx={{
+          width: setWidth(inputSize),
           variant:
             hasError && !hasFocus
               ? 'elements.simple-input.variants.error'
@@ -39,13 +67,17 @@ export const SimpleInput = React.forwardRef(
         }}
       >
         {prefix && (
-          <span sx={{ variant: 'elements.simple-input.affix.prefix' }}>
+          <span
+            ref={prefixRef}
+            sx={{ variant: 'elements.simple-input.affix.prefix' }}
+          >
             {prefix}
           </span>
         )}
         <input
           sx={{
-            variant: 'elements.simple-input.base'
+            variant: 'elements.simple-input.base',
+            textAlign: inputSize === 'sm' ? 'center' : 'left'
           }}
           placeholder={placeholder}
           ref={ref}
@@ -56,7 +88,10 @@ export const SimpleInput = React.forwardRef(
           {...props}
         />
         {suffix && (
-          <span sx={{ variant: 'elements.simple-input.affix.suffix' }}>
+          <span
+            ref={suffixRef}
+            sx={{ variant: 'elements.simple-input.affix.suffix' }}
+          >
             {suffix}
           </span>
         )}