+++ src/elements/cta/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.11](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.10...@uswitch/trustyle.cta@1.0.11) (2020-11-13)
+
+**Note:** Version bump only for package @uswitch/trustyle.cta
+
+
+
+
+
+## [1.0.10](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.9...@uswitch/trustyle.cta@1.0.10) (2020-11-11)
+
+**Note:** Version bump only for package @uswitch/trustyle.cta
+
+
+
+
+
+## [1.0.9](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.8...@uswitch/trustyle.cta@1.0.9) (2020-11-09)
+
+**Note:** Version bump only for package @uswitch/trustyle.cta
+
+
+
+
+
+## [1.0.8](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.7...@uswitch/trustyle.cta@1.0.8) (2020-11-06)
+
+**Note:** Version bump only for package @uswitch/trustyle.cta
+
+
+
+
+
+## [1.0.7](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.6...@uswitch/trustyle.cta@1.0.7) (2020-11-03)
+
+**Note:** Version bump only for package @uswitch/trustyle.cta
+
+
+
+
+
 ## [1.0.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.cta@1.0.1...@uswitch/trustyle.cta@1.0.2) (2020-08-11)
 
 **Note:** Version bump only for package @uswitch/trustyle.cta