+++ src/elements/embedded-video/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.33](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.32...@uswitch/trustyle.embedded-video@0.4.33) (2020-11-16)
+
+**Note:** Version bump only for package @uswitch/trustyle.embedded-video
+
+
+
+
+
+## [0.4.32](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.31...@uswitch/trustyle.embedded-video@0.4.32) (2020-11-16)
+
+**Note:** Version bump only for package @uswitch/trustyle.embedded-video
+
+
+
+
+
+## [0.4.31](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.30...@uswitch/trustyle.embedded-video@0.4.31) (2020-11-11)
+
+**Note:** Version bump only for package @uswitch/trustyle.embedded-video
+
+
+
+
+
+## [0.4.30](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.29...@uswitch/trustyle.embedded-video@0.4.30) (2020-11-09)
+
+**Note:** Version bump only for package @uswitch/trustyle.embedded-video
+
+
+
+
+
+## [0.4.29](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.28...@uswitch/trustyle.embedded-video@0.4.29) (2020-11-06)
+
+**Note:** Version bump only for package @uswitch/trustyle.embedded-video
+
+
+
+
+
 ## [0.4.27](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.embedded-video@0.4.26...@uswitch/trustyle.embedded-video@0.4.27) (2020-10-28)
 
 **Note:** Version bump only for package @uswitch/trustyle.embedded-video